requirements.txt
attrdict==2.0.1
pillow==8.0.1
requests==2.25.0